+++ src/main/java/com/sparta/todo/comment/dto/CommentRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.todo.comment.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class CommentRequestDto {
+
+    private String comment;
+}
