+++ src/main/java/com/sparta/todo/comment/dto/CommentResponseDto.java
@@ -1,0 +1,25 @@
+package com.sparta.todo.comment.dto;
+
+import com.sparta.todo.comment.entity.Comment;
+import java.time.LocalDateTime;
+import lombok.Getter;
+
+@Getter
+public class CommentResponseDto {
+
+    private Long id;
+    private String username;
+    private String comment;
+    private LocalDateTime createdAt;
+
+    public CommentResponseDto(Comment comment) {
+        this.id = comment.getId();
+        this.username = comment.getUsername();
+        this.comment = comment.getComment();
+        this.createdAt = comment.getCreatedAt();
+    }
+
+    public void update(CommentRequestDto requestDto) {
+        this.comment = requestDto.getComment();
+    }
+}
