+++ src/main/java/com/sparta/todo/comment/repository/CommentRepository.java
@@ -1,0 +1,10 @@
+package com.sparta.todo.comment.repository;
+
+import com.sparta.todo.comment.entity.Comment;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+
+    Optional<Comment> findByIdAndUserId(Long id, Long userId);
+}
