--- src/main/java/com/sparta/todo/dto/CommentRequestDto.java
@@ -1,10 +1,0 @@
-package com.sparta.todo.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CommentRequestDto {
-    private String comment;
-}
