--- src/main/java/com/sparta/todo/dto/CommentResponseDto.java
@@ -1,25 +1,0 @@
-package com.sparta.todo.dto;
-
-import com.sparta.todo.entity.Comment;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-public class CommentResponseDto {
-    private Long id;
-    private String username;
-    private String comment;
-    private LocalDateTime createdAt;
-
-    public CommentResponseDto(Comment comment) {
-        this.id = comment.getId();
-        this.username = comment.getUsername();
-        this.comment = comment.getComment();
-        this.createdAt = comment.getCreatedAt();
-    }
-
-    public void update(CommentRequestDto requestDto) {
-        this.comment = requestDto.getComment();
-    }
-}
