--- src/main/java/com/sparta/todo/dto/LoginRequestDto.java
@@ -1,12 +1,0 @@
-package com.sparta.todo.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@NoArgsConstructor
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
