--- src/main/java/com/sparta/todo/dto/SignupRequestDto.java
@@ -1,18 +1,0 @@
-package com.sparta.todo.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
-import lombok.Getter;
-
-@Getter
-public class SignupRequestDto {
-    @NotBlank
-    @Size(min = 4, max = 10, message = "최소 4자 이상, 10자 이하이며 알파벳 소문자(a~z), 숫자(0~9)")
-    @Pattern(regexp = "^[a-z0-9]{4,10}$")
-    private String username;
-
-    @NotBlank
-    @Pattern(regexp = "^[a-zA-Z0-9]{8,15}$", message = "최소 8자 이상, 15자 이하이며 알파벳 대소문자(a~z, A~Z), 숫자(0~9)")
-    private String password;
-}
