--- src/main/java/com/sparta/todo/dto/TodoRequestDto.java
@@ -1,10 +1,0 @@
-package com.sparta.todo.dto;
-
-import lombok.Getter;
-
-@Getter
-public class TodoRequestDto {
-    private String username;
-    private String title;
-    private String content;
-}
