--- src/main/java/com/sparta/todo/entity/Todo.java
@@ -1,60 +1,0 @@
-package com.sparta.todo.entity;
-
-import com.sparta.todo.dto.TodoRequestDto;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@Table(name = "todo")
-public class Todo extends Timestamped{
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "username", nullable = false)
-    private String username;
-
-    @Column(name = "title", nullable = false)
-    private String title;
-
-    @Column(name = "contents", nullable = false, length = 500)
-    private String content;
-
-    @Column(name = "completed", nullable = false)
-    private boolean isCompleted = false;
-
-
-    // Todo User 간의 다대일 관계 설정
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    // Todo Comment 간의 일대다 관계 설정
-    @OneToMany(mappedBy = "todo", cascade = CascadeType.PERSIST, orphanRemoval = true)
-    private List<Comment> comments = new ArrayList<>();
-
-    // todo 작성
-    public Todo (TodoRequestDto requestDto, User user) {
-        this.username = user.getUsername();
-        this.title = requestDto.getTitle();
-        this.content = requestDto.getContent();
-        this.isCompleted = isCompleted();
-        this.user = user;
-    }
-
-    // todo 수정
-    // 게시글 수정
-    public void update(TodoRequestDto requestDto) {
-        this.title = requestDto.getTitle();
-        this.content = requestDto.getContent();
-    }
-}
