--- src/main/java/com/sparta/todo/entity/UserRoleEnum.java
@@ -1,20 +1,0 @@
-package com.sparta.todo.entity;
-
-// 사용자 권한 관리 클래스
-public enum UserRoleEnum {
-    USER(Authority.USER); // 사용자 권한
-
-    private final String authority;
-
-    UserRoleEnum(String authority) {
-        this.authority = authority;
-    }
-
-    public String getAuthority() {
-        return this.authority;
-    }
-
-    public static class Authority {
-        public static final String USER = "ROLE_USER";
-    }
-}
