+++ src/main/java/com/sparta/todo/exception/ErrorResponse.java
@@ -1,0 +1,11 @@
+package com.sparta.todo.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorResponse {
+
+    private final String errorMessage;
+}
