+++ src/main/java/com/sparta/todo/filter/JwtAuthenticationFilter.java
@@ -1,0 +1,88 @@
+package com.sparta.todo.filter;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sparta.todo.exception.ResponseStatusDto;
+import com.sparta.todo.exception.Status;
+import com.sparta.todo.user.dto.LoginRequestDto;
+import com.sparta.todo.util.JwtUtil;
+import com.sparta.todo.util.UserDetailsImpl;
+import com.sparta.todo.util.UserRoleEnum;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+// filter단에서 인증 인가 처리
+@Slf4j(topic = "로그인 및 JWT 생성")
+public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+
+    private final JwtUtil jwtUtil;
+    private ObjectMapper objectMapper = new ObjectMapper();
+
+    public JwtAuthenticationFilter(JwtUtil jwtUtil) {
+        this.jwtUtil = jwtUtil;
+        setFilterProcessesUrl("/api/auth/login");
+    }
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request,
+            HttpServletResponse response) throws AuthenticationException {
+        log.info("로그인 시도");
+        try {
+            LoginRequestDto requestDto = new ObjectMapper().readValue(request.getInputStream(),
+                    LoginRequestDto.class);
+
+            return getAuthenticationManager().authenticate(
+                    new UsernamePasswordAuthenticationToken(
+                            requestDto.getUsername(),
+                            requestDto.getPassword(),
+                            null
+                    )
+            );
+        } catch (IOException e) {
+            log.error(e.getMessage());
+            throw new RuntimeException(e.getMessage());
+        }
+    }
+
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request,
+            HttpServletResponse response, FilterChain chain, Authentication authResult)
+            throws IOException, ServletException {
+        log.info("로그인 성공 및 JWT 생성");
+        response.setStatus(200); // 인증 성공
+        String username = ((UserDetailsImpl) authResult.getPrincipal()).getUsername();
+        UserRoleEnum role = ((UserDetailsImpl) authResult.getPrincipal()).getUser().getRole();
+
+        String token = jwtUtil.createToken(username, role);
+        jwtUtil.addJwtToCookie(token, response);
+
+        response.setContentType("application/json");
+        response.setCharacterEncoding("utf-8");
+
+        String result = objectMapper.writeValueAsString(
+                new ResponseStatusDto(Status.LOGIN_SUCCESS));
+        response.getWriter().write(result);
+    }
+
+    @Override
+    protected void unsuccessfulAuthentication(HttpServletRequest request,
+            HttpServletResponse response, AuthenticationException failed)
+            throws IOException, ServletException {
+        log.info("로그인 실패");
+        response.setStatus(401); // 인증 실패
+
+        response.setContentType("application/json");
+        response.setCharacterEncoding("utf-8");
+
+        String result = objectMapper.writeValueAsString(
+                new ResponseStatusDto(Status.LOGIN_FAILURE));
+        response.getWriter().write(result);
+    }
+}
