--- src/main/java/com/sparta/todo/repository/CommentRepository.java
@@ -1,10 +1,0 @@
-package com.sparta.todo.repository;
-
-import com.sparta.todo.entity.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    Optional<Comment> findByIdAndUserId(Long id, Long userId);
-}
