--- src/main/java/com/sparta/todo/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.sparta.todo.repository;
-
-import com.sparta.todo.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
