--- src/main/java/com/sparta/todo/service/UserService.java
@@ -1,41 +1,0 @@
-package com.sparta.todo.service;
-
-import com.sparta.todo.dto.SignupRequestDto;
-import com.sparta.todo.entity.User;
-import com.sparta.todo.entity.UserRoleEnum;
-import com.sparta.todo.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public void signup(SignupRequestDto requestDto) {
-
-        String username = requestDto.getUsername();
-        String password = passwordEncoder.encode(requestDto.getPassword());
-
-        // 회원 중복 확인
-        Optional<User> checkUsername = userRepository.findByUsername(username);
-        if (checkUsername.isPresent()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "중복된 username 입니다.");
-        }
-
-        // 사용자 ROLE 확인
-        UserRoleEnum role = UserRoleEnum.USER;
-
-        // 사용자 등록
-        User user = new User(username, password, role);
-        userRepository.save(user);
-    }
-
-}
