+++ src/main/java/com/sparta/todo/todo/dto/TodoRequestDto.java
@@ -1,0 +1,11 @@
+package com.sparta.todo.todo.dto;
+
+import lombok.Getter;
+
+@Getter
+public class TodoRequestDto {
+
+    private String username;
+    private String title;
+    private String content;
+}
