+++ src/main/java/com/sparta/todo/user/dto/LoginRequestDto.java
@@ -1,0 +1,12 @@
+package com.sparta.todo.user.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class LoginRequestDto {
+
+    private String username;
+    private String password;
+}
