+++ src/main/java/com/sparta/todo/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.sparta.todo.user.repository;
+
+import com.sparta.todo.user.entity.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByUsername(String username);
+}
