+++ src/main/java/com/sparta/todo/user/service/UserService.java
@@ -1,0 +1,39 @@
+package com.sparta.todo.user.service;
+
+import com.sparta.todo.user.dto.SignupRequestDto;
+import com.sparta.todo.user.entity.User;
+import com.sparta.todo.user.repository.UserRepository;
+import com.sparta.todo.util.UserRoleEnum;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    public void signup(SignupRequestDto requestDto) {
+
+        String username = requestDto.getUsername();
+        String password = passwordEncoder.encode(requestDto.getPassword());
+
+        // 회원 중복 확인
+        Optional<User> checkUsername = userRepository.findByUsername(username);
+        if (checkUsername.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "중복된 username 입니다.");
+        }
+
+        // 사용자 ROLE 확인
+        UserRoleEnum role = UserRoleEnum.USER;
+
+        // 사용자 등록
+        User user = new User(username, password, role);
+        userRepository.save(user);
+    }
+}
